// pronti-frontend-main/front/src/app/modulos/cargo-vacacion/servicios/dto/cargo-vacacion.update-dto.ts
export interface CargoVacacionUpdateDto {

  // nombre?: string;
  fechaDesde?: string;
  fechaHasta?: string;
  diasVacaciones?: number;
  diasAntiguedad?: number;
  diasTeoricos?: number;
  diasTomados?: number;
  diasSaldo?: number;
  valorVacaciones?: number;
  valorDias?: number;
  valorAntiguedad?: number;
  valorTeorico?: number;
  valorTomado?: number;
  valorSaldo?: number;
  totalIngresosAnio?: number;
  numAnioAcumulado?: number;
}